refactor: 번호 생성 로직을 util 패키지로 이동
