+++ src/main/java/util/LottoNumbersGenerator.java
@@ -1,4 +1,4 @@
-package domain;
+package util;
 
 import java.util.ArrayList;
 import java.util.Collections;
