+++ src/main/java/util/NumbersGenerator.java
@@ -1,4 +1,4 @@
-package domain;
+package util;
 
 import java.util.List;
